Move org constants to constants file
And create constants out of the individual org types.
