+++ app/organisation/rest.py
@@ -2,7 +2,7 @@ from flask import Blueprint, abort, current_app, jsonify, request
 from sqlalchemy.exc import IntegrityError
 
 from app.config import QueueNames
-from app.constants import INVITE_PENDING, KEY_TYPE_NORMAL
+from app.constants import INVITE_PENDING, KEY_TYPE_NORMAL, NHS_ORGANISATION_TYPES
 from app.dao.annual_billing_dao import set_default_free_allowance_for_service
 from app.dao.dao_utils import transaction
 from app.dao.fact_billing_dao import fetch_usage_for_organisation
@@ -30,7 +30,7 @@ from app.dao.services_dao import dao_fetch_service_by_id
 from app.dao.templates_dao import dao_get_template_by_id
 from app.dao.users_dao import get_user_by_id
 from app.errors import InvalidRequest, register_errors
-from app.models import NHS_ORGANISATION_TYPES, Organisation
+from app.models import Organisation
 from app.notifications.process_notifications import (
     persist_notification,
     send_notification_to_queue,
